requirements.txt
numpy
requests
torch
torchvision
pandas
Pillow
patool